SType replaces explicit size_t and int
(1) SType is now template parameter that replaces size_t in many places
(2) All class template expose their FType and SType as F & S typedefs
(3) If needed classes use SS typedef for signed type of S

All these changes are done to resolve unsigned <-> signed conversions and type
analysis warnings.

+++ src/pet_matrix.cpp
@@ -33,13 +33,12 @@ int main(int argc, char* argv[]) {
     cl.add<cmdline::string>(
         "config", 'c', "load config file", false, cmdline::string(), false);
 #if _OPENMP
-    cl.add<size_t>(
-        "n-threads", 't', "number of OpenMP threads", false, 0, false);
+    cl.add<int>("n-threads", 't', "number of OpenMP threads", false, 0, false);
 #endif
-    cl.add<size_t>(
+    cl.add<int>(
         "n-pixels", 'n', "number of pixels in one dimension", false, 256);
-    cl.add<size_t>("n-detectors", 'd', "number of ring detectors", false, 64);
-    cl.add<size_t>("n-emissions", 'e', "emissions per pixel", false, 0);
+    cl.add<int>("n-detectors", 'd', "number of ring detectors", false, 64);
+    cl.add<int>("n-emissions", 'e', "emissions per pixel", false, 0);
     cl.add<double>("radius", 'r', "inner detector ring radius", false, 0);
     cl.add<double>("s-pixel", 'p', "pixel size", false);
     cl.add<double>("w-detector", 'w', "detector width", false);
@@ -66,9 +65,8 @@ int main(int argc, char* argv[]) {
     // visual debugging params
     cl.add<cmdline::string>(
         "png", 0, "output lor to png", false, cmdline::string(), false);
-    cl.add<ssize_t>(
-        "from", 0, "lor start detector to output", false, -1, false);
-    cl.add<ssize_t>("to", 0, "lor end detector to output", false, -1, false);
+    cl.add<int>("from", 0, "lor start detector to output", false, -1, false);
+    cl.add<int>("to", 0, "lor end detector to output", false, -1, false);
 
     // printing & stats params
     cl.add("print", 0, "print triangular sparse system matrix");
@@ -77,9 +75,9 @@ int main(int argc, char* argv[]) {
 
     cl.parse_check(argc, argv);
 
-    auto& n_pixels = cl.get<size_t>("n-pixels");
-    auto& n_detectors = cl.get<size_t>("n-detectors");
-    auto& n_emissions = cl.get<size_t>("n-emissions");
+    auto& n_pixels = cl.get<int>("n-pixels");
+    auto& n_detectors = cl.get<int>("n-detectors");
+    auto& n_emissions = cl.get<int>("n-emissions");
     auto& radius = cl.get<double>("radius");
     auto& s_pixel = cl.get<double>("s-pixel");
     auto& w_detector = cl.get<double>("w-detector");
@@ -126,7 +124,7 @@ int main(int argc, char* argv[]) {
 
 #if _OPENMP
     if (cl.exist("n-threads")) {
-      omp_set_num_threads(cl.get<size_t>("n-threads"));
+      omp_set_num_threads(cl.get<int>("n-threads"));
     }
 #endif
 
@@ -232,9 +230,9 @@ int main(int argc, char* argv[]) {
     // visual debugging output
     if (cl.exist("png")) {
       MatrixLORMajor<>::LOR lor(0, 0);
-      lor.first = cl.get<ssize_t>("from");
+      lor.first = cl.get<int>("from");
       if (cl.exist("to")) {
-        lor.second = cl.get<ssize_t>("to");
+        lor.second = cl.get<int>("to");
       } else {
         lor.second = (lor.first + n_detectors / 2) % n_detectors;
       }
